harmonisation et correction

+++ src/controllers/roles.controller.ts
@@ -7,16 +7,18 @@ const rolesController = {
     get: async (request: Request, response: Response) => {
         try {
             const { id } = request.params;
-            logger.info("[GET] Récupérer une role"); // Log d'information en couleur
+
+            logger.info(`[GET] Récupérer le role avec l'id: ${id}`);
+
             const role = await rolesModel.get(id);
             if (!role) {
+                logger.error("Role inexistant");
                 return APIResponse(response, null, "Role inexistant", 404);
             }
+
             APIResponse(response, role, "OK");
         } catch (error: any) {
-            logger.error(
-                "Erreur lors de la récupération du role: " + error.message,
-            );
+            logger.error("Erreur lors de la récupération du role: ", error);
             APIResponse(
                 response,
                 null,
@@ -31,16 +33,14 @@ const rolesController = {
                 name
             } = rolesRegisterValidation.parse(request.body);
 
-            logger.info("[POST] Créer une role"); // Log d'information en couleur
+            logger.info("[POST] Créer un role"); // Log d'information en couleur
 
             const role = await rolesModel.create({
                 name
             });
             APIResponse(response, role, "OK", 201);
         } catch (error: any) {
-            logger.error(
-                "Erreur lors de la récupération du role: " + error.message,
-            );
+            logger.error("Erreur lors de la récupération du role: ", error);
             APIResponse(
                 response,
                 null,
@@ -53,14 +53,18 @@ const rolesController = {
         try {
             const { id } = request.params;
 
-            logger.info("[DELETE] Supprimer une role"); // Log d'information en couleur
+            logger.info(`[DELETE] Supprimer le role avec l'id: ${id}`);
+
+            const role = await rolesModel.get(id);
+            if (!role) {
+                logger.error("Role inexistant");
+                return APIResponse(response, null, "Role inexistant", 404);
+            }
             
             await rolesModel.delete(id);
             APIResponse(response, null, "OK", 201);
         } catch (error: any) {
-            logger.error(
-                "Erreur lors de la suppression du role: " + error.message,
-            );
+            logger.error("Erreur lors de la suppression du role: ", error);
             APIResponse(
                 response,
                 null,
@@ -72,18 +76,23 @@ const rolesController = {
     update: async (request: Request, response: Response) => {
         try {
             const { id } = request.params;
-            const {
-                name
-            } = rolesRegisterValidation.parse(request.body);
-            
-            logger.info("[UPDATE] Update une role"); // Log d'information en couleur
+
+            logger.info(`[UPDATE] Modifier le role avec l'id: ${id}`);
+
+            const role = await rolesModel.get(id);
+            if (!role) {
+                logger.error("Role inexistant");
+                return APIResponse(response, null, "Role inexistant", 404);
+            }
+
+            const { name } = rolesRegisterValidation.parse(request.body);
             
             await rolesModel.update(id, {
                 name
             });
             APIResponse(response, null, "OK", 201);
         } catch (error: any) {
-            logger.error("Erreur lors de la màj du role: " + error.message);
+            logger.error("Erreur lors de la màj du role: ", error);
             APIResponse(
                 response,
                 null,
@@ -99,10 +108,7 @@ const rolesController = {
             const roles = await rolesModel.getAll();
             APIResponse(response, roles, "OK");
         } catch (error: any) {
-            logger.error(
-                "Erreur lors de la récupération des roles: " +
-                    error.message,
-            );
+            logger.error("Erreur lors de la récupération des roles: ", error);
             APIResponse(
                 response,
                 null,
